extract methods to trait
